adds selectors
