+++ client/js/_author/selectors/assessment.js
@@ -1,0 +1,11 @@
+import { createSelector }  from 'reselect';
+
+export function transformAssessment(assessment) {
+  if (!assessment) return {};
+  const fixedAssessment = {
+    ...assessment,
+    name: assessment.displayName.text,
+  };
+
+  return fixedAssessment;
+}
